ModularPP and FracGrad changed to hoopmax
This is part 1 of 2, the long-standing problem of assuming the fracgrad is equal to SHmin is kinda solved, but not quite. While Zhang in his book refers to something like using hoop stress to compute frac grad, I will check with the text to see if we implemented it correctly, but this is still not quite right.

True frac grad is the delta increase in mudweight that causes the minimum hoopstress to exceed the tensile strength of the rock. I need a way to figure out how best to compute this, as well as to account for thermal effects in the stress calculation along wellbore