Bugfixes
Recent changes being ironed out to be compatible with most las files and situations.
The stresses are calculated at the same resolution as everything else, there is an experimental fix for length matching in getwelldev, and a bugfix in the initial downsampler implementation. Also wellisvertical is nerfed to being a vestigal function now, all deviation processing being done in getwelldev in a unified manner, whether or not well is vertical