bugfixes second round
Fix the length not matching, better technique this time, not relying on las header data which may be incorrect/unavailable.
Put the dexp in a try block, there seemed to be some issue with the function, was not obvious what it was.
Up next, the averaging function needs to be looked at (the median one seems fine kinda?)